Add files via upload
This Project is very useful for people to find the information related to trains based on their origin and destination stations. Users upon entering their origin and destination and also the day of the week they are planning to travel will get a whole list of trains available based on their requirement. They will also be getting the feature of providing their choice of coach that they would like to travel, and in-return will be provided with the number of available seats and their fare which would help them to compare against the other train and choose the most optimal train for them. This system can be also integrated with a booking system and an end-to-end application can be created out of this. 

This can also be integrated with various 3rd party vendors such as Goibibo, MakeMyTrip etc., to provide their users with this functionality where they can quickly look for the available trains and they don’t have to login every time to the Railway Website to fetch these details. The Purpose of this project is to provide a seamless and hassle-free system for the users to view their choice of trains.

Requirements: Python3, Jupyter Notebook, Processors: i3, Disk space: 2 to 4 GB.